% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_func.R
\name{get_fcc}
\alias{get_fcc}
\title{Fold-change concordance (FCC) score}
\usage{
get_fcc(fc_vect)
}
\arguments{
\item{fc_vect}{Vector of fold-changes.}
}
\value{
The FCC score.
}
\description{
Function that calculates the FCC score for a vector of fold-change (FC) values.
}
