% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_func.R
\name{get_ratioDown}
\alias{get_ratioDown}
\title{RatioDown score}
\usage{
get_ratioDown(fc_vect)
}
\arguments{
\item{fc_vect}{Vector of fold-changes.}
}
\value{
The ratioDown score.
}
\description{
Function that calculates ratio of down-regulated genes for a vector of fold-change (FC) values.
}
