% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scores_func.R
\name{stouffer}
\alias{stouffer}
\title{Stouffer's method for p-values combination}
\usage{
stouffer(ps, two.tails = FALSE)
}
\arguments{
\item{ps}{The p-values to be combined.}
}
\value{
The normalized dataframe.
}
\description{
Function that combines p-values using Stouffer's method.
}
